requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Cerberus==1.3.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.6
django-allauth==0.51.0
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-modal-forms==2.2.0
django-bootstrap4==22.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-widget-tweaks==1.4.12
fontawesomefree==6.5.1
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
static3==0.7.0
urllib3==1.26.10
whitenoise==6.2.0
